thor: Implement gic abstraction
