+++ kernel/thor/arch/arm/meson.build
@@ -6,6 +6,7 @@ src += files(
 	'cpu.cpp',
 	'debug.cpp',
 	'gic.cpp',
+	'gic_v2.cpp',
 	'ints.cpp',
 	'paging.cpp',
 	'system.cpp',
