+++ kernel/thor/system/dtb/dtb.cpp
@@ -393,7 +393,7 @@ auto DeviceTreeNode::parseIrqs_(frg::span<const std::byte> data) -> frg::vector<
 	::DeviceTreeProperty prop{"", data};
 
 	// We only support GIC irqs for now
-	if (!isCompatible(dtGicCompatible)) {
+	if (!isCompatible(dtGicV2Compatible)) {
 		infoLogger() << "thor: warning: Skipping parsing IRQs using node \"" << path()
 			<< "\", it's not compatible with the GIC" << frg::endlog;
 		return ret;
